// crm_frontend_3/src/api/api3.ts
export class DeleteClient extends ClientBase {
    private http: { fetch(url: RequestInfo, init?: RequestInit): Promise<Response> };
    private baseUrl: string;
    protected jsonParseReviver: ((key: string, value: any) => any) | undefined = undefined;

    constructor(baseUrl?: string, http?: { fetch(url: RequestInfo, init?: RequestInit): Promise<Response> }) {
        super();
        this.http = http ? http : window as any;
        this.baseUrl = baseUrl !== undefined && baseUrl !== null ? baseUrl : "";
    }

    /**
     * @param id_accessories (optional) 
     * @param id_order (optional) 
     * @return Success
     */
     delet_equipment(version: string, body: Delet_EquipmentDto): Promise<void> {
        let url_ = this.baseUrl + "/api/{version}/Storekeeper/equipment/delet";
        if (version === undefined || version === null)
            throw new Error("The parameter 'version' must be defined.");
        url_ = url_.replace("{version}", encodeURIComponent("" + version));
        url_ = url_.replace(/[?&]$/, "");

        const content_ = JSON.stringify(body);

        let options_: RequestInit = {
            body: content_,
            method: "DELETE",
            headers: {
                "Content-Type": "application/json",
            }
        };

        return this.transformOptions(options_).then(transformedOptions_ => {
            return this.http.fetch(url_, transformedOptions_);
        }).then((_response: Response) => {
            return this.processEquipment(_response);
        });
    }

    protected processEquipment(response: Response): Promise<void> {
        const status = response.status;
        let _headers: any = {}; if (response.headers && response.headers.forEach) { response.headers.forEach((v: any, k: any) => _headers[k] = v); };
        if (status === 204) {
            return response.text().then((_responseText) => {
            return;
            });
        } else if (status === 401) {
            return response.text().then((_responseText) => {
            let result401: any = null;
            result401 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver) as ProblemDetails;
            return throwException("Unauthorized", status, _responseText, _headers, result401);
            });
        } else if (status !== 200 && status !== 204) {
            return response.text().then((_responseText) => {
            return throwException("An unexpected server error occurred.", status, _responseText, _headers);
            });
        }
        return Promise.resolve<void>(null as any);
    }
    /**
     * @param body (optional) 
     * @return Success
     */
     delet_services(version: string, body: Delet_ServicesDto ): Promise<void> {
        let url_ = this.baseUrl + "/api/{version}/Storekeeper/services/delet";
        if (version === undefined || version === null)
            throw new Error("The parameter 'version' must be defined.");
        url_ = url_.replace("{version}", encodeURIComponent("" + version));
        url_ = url_.replace(/[?&]$/, "");

        const content_ = JSON.stringify(body);

        let options_: RequestInit = {
            body: content_,
            method: "DELETE",
            headers: {
                "Content-Type": "application/json",
            }
        };

        return this.transformOptions(options_).then(transformedOptions_ => {
            return this.http.fetch(url_, transformedOptions_);
        }).then((_response: Response) => {
            return this.processServices(_response);
        });
    }

    protected processServices(response: Response): Promise<void> {
        const status = response.status;
        let _headers: any = {}; if (response.headers && response.headers.forEach) { response.headers.forEach((v: any, k: any) => _headers[k] = v); };
        if (status === 204) {
            return response.text().then((_responseText) => {
            return;
            });
        } else if (status === 401) {
            return response.text().then((_responseText) => {
            let result401: any = null;
            result401 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver) as ProblemDetails;
            return throwException("Unauthorized", status, _responseText, _headers, result401);
            });
        } else if (status !== 200 && status !== 204) {
            return response.text().then((_responseText) => {
            return throwException("An unexpected server error occurred.", status, _responseText, _headers);
            });
        }
        return Promise.resolve<void>(null as any);
    }
}

export interface Accessories_CPUVm {
    iD_Сomponent: string;
    name_Сomponent: string ;
    price_CPU: number;
    quantity_CPU: number;
    number_Cores: number;
    purity_CPU: string ;
}

export interface Creat_EquimpentDto {
    iD_Order: number;
    iD_Accessories: number;
    name_Company: string ;
}

export interface List_Equipment {
    iD_Order: number;
    iD_Accessories: number;
    name_Сomponent: string ;
    type_Сomponent: string ;
    price_Сomponent: number;
    name_Company: string;
}

export interface List_EquipmentVm {
    list_Equipments: List_Equipment[] ;
}

export interface List_Master_ClientVm {
    orders: List_Order_Master[] ;
}

export interface Delet_EquipmentDto {
    iD_Order: number;
    iD_Accessories: number;
}
// ...
export interface Delet_ServicesDto {
    iD_Services: number;
}


export interface List_Services {
    iD_Services: number;
    iD_Order: number;
    name: string ;
    description: string;
    price_services: number;
}

export interface List_ServicesVm {
    list_Servicess: List_Services[] ;
}


export interface ProblemDetails {
    type?: string | undefined;
    title?: string | undefined;
    status?: number | undefined;
    detail?: string | undefined;
    instance?: string | undefined;
}